.gitignore changed

+++ src/main/java/com/agazibaric/entity/Item.java
@@ -5,14 +5,14 @@ import java.time.LocalDateTime;
 
 public class Item {
 
-    private int id;
+    private long id;
     private String name;
     private String description;
     private float price;
     private LocalDateTime creationTime;
     private Duration duration;
 
-    public Item(int id, String name, String description, float price, LocalDateTime creationTime, Duration duration) {
+    public Item(long id, String name, String description, float price, LocalDateTime creationTime, Duration duration) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -23,11 +23,11 @@ public class Item {
 
     public Item() {}
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
